backend_api/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.1
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==23.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
Faker==36.1.1
flower==2.0.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
numpy==2.2.5
oauthlib==3.2.2
prometheus_client==0.22.1
prompt_toolkit==3.0.50
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.3
sendgrid==6.11.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
threadpoolctl==3.6.0
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
